requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
stripe==2.59.0
python-dotenv==1.0.0
sendgrid==6.9.7